chore(server-split): v0.4.x PR-C — jobs/scheduler routes moved to routes/jobs.py (9 endpoints + 3 Pydantic + DI lazy forwarder)
docs/design/v0.4.x-server-split.md 8-PR sequence 중 네 번째 PR.
PR-A.0 + PR-A.1 + PR-B 위에 W8 workspace jobs + scheduler 분리.

## 이동된 surface (253 lines 서버에서 제거)

### 9 endpoints
- /jobs/run, /jobs/schedule, /jobs/unschedule
- /jobs/list, /jobs/{job_id}, /jobs/{job_id}/download
- /admin/scheduler/status
- /admin/jobs/list, /admin/jobs/{job_id}

### 3 Pydantic 모델
- JobRunRequest, JobScheduleRequest, JobUnscheduleRequest

## routes/_deps.py — DI lazy-forwarder 패턴 전환

기존 PR-A.0 의 snapshot-via-set_* 패턴은 monkeypatch 호환성 깨짐:
- 테스트가 `srv.rag_engine = stub` 로 attribute 패치 → server module 만 변경
- routes/_deps._rag_engine 은 boot 시점 snapshot 이라 stub 안 봄
- 결과: extracted router 가 stub 대신 원본 engine 사용

전환:
```python
def get_rag_engine():
    import server_llmwiki  # lazy — 호출 시점, circular 회피
    return server_llmwiki.rag_engine
```

- 모든 get_* 가 lazy-forwarder 로 변환
- set_* 는 no-op 으로 유지 (PR-A.0 의 server boot 코드와 back-compat)
- 단일 SOT (server module attribute) + monkeypatch 자연 통과

검증: tests/test_workspace_jobs.py + tests/test_scheduler.py 의
`srv.rag_engine = _StubEngine()` 패턴 모두 그대로 작동.

## tests/_server_split_helpers.py 갱신

`combined_server_source()` 의 routes/* 모듈 목록에 `routes.jobs` 추가.

## server_llmwiki.py 변경

- 9 endpoint + 3 Pydantic 제거 (~253 lines)
- `from routes.jobs import router as jobs_router` + `app.include_router(jobs_router)` (llm_router 다음)
- ruff --fix 가 unused import 자동 정리

## 검증 (design memo §5 contract)

| invariant | 결과 |
|---|---|
| 1. URL byte-identical | ✓ `python scripts/audit_endpoint_paths.py origin/main` → **OK: 125 endpoints identical to origin/main** |
| 2. Middleware order | ✓ test_middleware_stack_includes_rate_limit_and_no_cache pass |
| 3. _write_audit emit | ✓ 24 사이트 동일 (routes/_helpers SOT) |
| 4. /static mount | ✓ test_static_mount_present pass |
| 5. startup event | ✓ test_startup_event_registered pass |
| 6. pytest 회귀 | ✓ **3294 passed, 0 failed** (2 .env-flakes deselected — pre-existing) |
| ruff F-class | ✓ All checks passed |

## 변경 통계

```
 routes/_deps.py                |  73 +++++-----
 routes/jobs.py                 | 302 +++++++++++++++++++++++++++++++++++++++++
 server_llmwiki.py              | 261 +----------------------------------
 tests/_server_split_helpers.py |   4 +-
 4 files changed, 345 insertions(+), 295 deletions(-)
```

server_llmwiki.py: 4685 → 4439 lines (-246).
8-PR roadmap 진행률 ≈ **25%** ((724+485+246)/5894 = 1455/5894).

## Out of scope (다음 PR 들로)

- PR-D: /artifacts/* + /admin/artifacts/* (W7-A data artifacts)
- PR-E: Phase 7 자기진화 + 멀티모달
- PR-F: /code/* (코딩 에이전트)
- PR-G: 잔여 /admin/*
- PR-H: server_llmwiki.py 최종 정리 (~500 lines target)

Quality delta: exempt (label: chore). URL 표면 byte-identical, core/ 무관.
